+++ examples/schunk_wsg/BUILD.bazel
@@ -33,8 +33,8 @@ drake_cc_binary(
         ":simulated_schunk_wsg_system",
         "//lcm",
         "//lcmtypes:schunk",
+        "//manipulation/schunk_wsg:lcm_schunk_wsg_controller",
         "//manipulation/schunk_wsg:schunk_wsg_constants",
-        "//manipulation/schunk_wsg:schunk_wsg_controller",
         "//manipulation/schunk_wsg:schunk_wsg_lcm",
         "//multibody/rigid_body_plant",
         "//multibody/rigid_body_plant:drake_visualizer",
@@ -63,8 +63,8 @@ drake_cc_googletest(
         "//common/trajectories:piecewise_polynomial_trajectory",
         "//common/trajectories:trajectory",
         "//lcm",
-        "//manipulation/schunk_wsg:schunk_wsg_force_controller",
         "//manipulation/schunk_wsg:schunk_wsg_constants",
+        "//manipulation/schunk_wsg:schunk_wsg_low_level_controller",
         "//multibody:rigid_body_tree",
         "//multibody:rigid_body_tree_construction",
         "//multibody/parsers",
