+++ manipulation/schunk_wsg/BUILD.bazel
@@ -43,25 +43,9 @@ drake_cc_library(
 )
 
 drake_cc_library(
-    name = "schunk_wsg_plain_controller",
-    srcs = ["schunk_wsg_plain_controller.cc"],
-    hdrs = ["schunk_wsg_plain_controller.h"],
-    deps = [
-        ":schunk_wsg_constants",
-        ":schunk_wsg_lcm",
-        ":schunk_wsg_position_controller",
-        "//drake/common:essential",
-        "//drake/systems/controllers:pid_controller",
-        "//drake/systems/framework:diagram",
-        "//drake/systems/framework:diagram_builder",
-        "//drake/systems/primitives",
-    ],
-)
-
-drake_cc_library(
-    name = "schunk_wsg_force_controller",
-    srcs = ["schunk_wsg_force_controller.cc"],
-    hdrs = ["schunk_wsg_force_controller.h"],
+    name = "schunk_wsg_low_level_controller",
+    srcs = ["schunk_wsg_low_level_controller.cc"],
+    hdrs = ["schunk_wsg_low_level_controller.h"],
     deps = [
         ":schunk_wsg_constants",
         ":schunk_wsg_lcm",
@@ -78,7 +62,7 @@ drake_cc_library(
     srcs = ["schunk_wsg_position_controller.cc"],
     hdrs = ["schunk_wsg_position_controller.h"],
     deps = [
-        ":schunk_wsg_force_controller",
+        ":schunk_wsg_low_level_controller",
     ],
 )
 
@@ -87,7 +71,23 @@ drake_cc_library(
     srcs = ["schunk_wsg_controller.cc"],
     hdrs = ["schunk_wsg_controller.h"],
     deps = [
-        ":schunk_wsg_plain_controller",
+        ":schunk_wsg_constants",
+        ":schunk_wsg_lcm",
+        ":schunk_wsg_position_controller",
+        "//drake/common:essential",
+        "//drake/systems/controllers:pid_controller",
+        "//drake/systems/framework:diagram",
+        "//drake/systems/framework:diagram_builder",
+        "//drake/systems/primitives",
+    ],
+)
+
+drake_cc_library(
+    name = "lcm_schunk_wsg_controller",
+    srcs = ["lcm_schunk_wsg_controller.cc"],
+    hdrs = ["lcm_schunk_wsg_controller.h"],
+    deps = [
+        ":schunk_wsg_controller",
     ],
 )
 
@@ -108,10 +108,10 @@ drake_cc_googletest(
 )
 
 drake_cc_googletest(
-    name = "schunk_wsg_controller_test",
+    name = "lcm_schunk_wsg_controller_test",
     deps = [
         ":schunk_wsg_constants",
-        ":schunk_wsg_controller",
+        ":lcm_schunk_wsg_controller",
         "//systems/analysis:simulator",
     ],
 )
