+++ manipulation/schunk_wsg/schunk_wsg_controller.cc
@@ -1,10 +1,17 @@
 #include "drake/manipulation/schunk_wsg/schunk_wsg_controller.h"
 
-#include "drake/manipulation/schunk_wsg/schunk_wsg_constants.h"
-#include "drake/manipulation/schunk_wsg/schunk_wsg_plain_controller.h"
-#include "drake/manipulation/schunk_wsg/schunk_wsg_lcm.h"
+#include <vector>
+
+#include "drake/manipulation/schunk_wsg/schunk_wsg_low_level_controller.h"
+#include "drake/manipulation/schunk_wsg/schunk_wsg_position_controller.h"
+#include "drake/systems/controllers/pid_controller.h"
 #include "drake/systems/framework/diagram_builder.h"
+#include "drake/systems/primitives/adder.h"
+#include "drake/systems/primitives/constant_vector_source.h"
+#include "drake/systems/primitives/matrix_gain.h"
+#include "drake/systems/primitives/multiplexer.h"
 #include "drake/systems/primitives/pass_through.h"
+#include "drake/systems/primitives/saturation.h"
 
 namespace drake {
 namespace manipulation {
@@ -13,37 +20,33 @@ namespace schunk_wsg {
 SchunkWsgController::SchunkWsgController() {
   systems::DiagramBuilder<double> builder;
 
-  auto state_pass_through = builder.AddSystem<systems::PassThrough<double>>(
-      kSchunkWsgNumPositions + kSchunkWsgNumVelocities);
+  // Add low-level controller. This makes the gripper behave as though it had
+  // only one degree of freedom.
+  auto low_level_controller = builder.AddSystem<SchunkWsgLowLevelController>();
 
-  auto wsg_trajectory_generator =
-      builder.AddSystem<SchunkWsgTrajectoryGenerator>(
-          kSchunkWsgNumPositions + kSchunkWsgNumVelocities,
-          kSchunkWsgPositionIndex);
-
-  // Add the LCM-free controller.
-  auto wsg_controller = builder.AddSystem<SchunkWsgPlainController>();
+  // Add position controller. This controls the separation of the fingers.
+  auto position_controller = builder.AddSystem<SchunkWsgPositionController>();
 
   // Export the inputs.
-  command_input_port_ =
-      builder.ExportInput(wsg_trajectory_generator->get_command_input_port());
-  state_input_port_ = builder.ExportInput(state_pass_through->get_input_port());
+  estimated_state_input_port_ = builder.ExportInput(
+      low_level_controller->get_estimated_joint_state_input_port());
+  desired_state_input_port_ =
+      builder.ExportInput(position_controller->get_input_port_desired_state());
+  max_force_input_port_ =
+      builder.ExportInput(position_controller->get_max_force_input_port());
 
   // Export the outputs.
-  builder.ExportOutput(wsg_controller->get_output_port_control());
+  builder.ExportOutput(
+      low_level_controller->get_comanded_joint_force_output_port());
 
   // Connect the subsystems.
-  // state_state_pass_through -> other subsystems
-  builder.Connect(state_pass_through->get_output_port(),
-                  wsg_trajectory_generator->get_state_input_port());
-  builder.Connect(state_pass_through->get_output_port(),
-                  wsg_controller->get_input_port_estimated_state());
+  // position_controller -> low_level_controller
+  builder.Connect(position_controller->get_output_port_control(),
+                  low_level_controller->get_commanded_grip_force_input_port());
 
-  // wsg_trajectory_generator -> position_controller
-  builder.Connect(wsg_trajectory_generator->get_target_output_port(),
-                  wsg_controller->get_input_port_desired_state());
-  builder.Connect(wsg_trajectory_generator->get_max_force_output_port(),
-                  wsg_controller->get_max_force_input_port());
+  // low_level_controller -> position_controller
+  builder.Connect(low_level_controller->get_estimated_grip_state_output_port(),
+                  position_controller->get_input_port_estimated_state());
 
   builder.BuildInto(this);
 }
