+++ manipulation/schunk_wsg/schunk_wsg_position_controller.cc
@@ -3,7 +3,6 @@
 #include <vector>
 
 #include "drake/manipulation/schunk_wsg/schunk_wsg_constants.h"
-#include "drake/manipulation/schunk_wsg/schunk_wsg_force_controller.h"
 #include "drake/systems/controllers/pid_controller.h"
 #include "drake/systems/framework/diagram_builder.h"
 #include "drake/systems/primitives/adder.h"
